pyproject.toml
[project]
name = "whatsapp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-neo4j>=0.8.0",
    "mem0ai>=1.0.5",
    "neo4j>=6.1.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
